Move application to lib
