+++ .gitignore
@@ -1,2 +1,3 @@
 node_modules/
 selenium-server-standalone-2.47.0.jar
+bin
